+++ cli/src/merge_tools/mod.rs
@@ -20,6 +20,7 @@ use std::sync::Arc;
 
 use bstr::BString;
 use itertools::Itertools as _;
+use jj_lib::backend::BackendError;
 use jj_lib::backend::FileId;
 use jj_lib::backend::MergedTreeId;
 use jj_lib::config::ConfigGetError;
@@ -32,6 +33,7 @@ use jj_lib::matchers::Matcher;
 use jj_lib::merge::Merge;
 use jj_lib::merge::MergedTreeValue;
 use jj_lib::merged_tree::MergedTree;
+use jj_lib::merged_tree::MergedTreeBuilder;
 use jj_lib::repo_path::InvalidRepoPathError;
 use jj_lib::repo_path::RepoPath;
 use jj_lib::repo_path::RepoPathBuf;
@@ -56,6 +58,8 @@ use crate::config::CommandNameAndArgs;
 use crate::ui::Ui;
 
 const BUILTIN_EDITOR_NAME: &str = ":builtin";
+const OURS_TOOL_NAME: &str = ":ours";
+const THEIRS_TOOL_NAME: &str = ":theirs";
 
 #[derive(Debug, Error)]
 pub enum DiffEditError {
@@ -127,6 +131,8 @@ pub enum MergeToolConfigError {
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub enum MergeTool {
     Builtin,
+    Ours,
+    Theirs,
     // Boxed because ExternalMergeTool is big compared to the Builtin variant.
     External(Box<ExternalMergeTool>),
 }
@@ -144,6 +150,8 @@ impl MergeTool {
     ) -> Result<Option<Self>, MergeToolConfigError> {
         match name {
             BUILTIN_EDITOR_NAME => Ok(Some(MergeTool::Builtin)),
+            OURS_TOOL_NAME => Ok(Some(MergeTool::Ours)),
+            THEIRS_TOOL_NAME => Ok(Some(MergeTool::Theirs)),
             _ => Ok(get_external_tool_config(settings, name)?.map(MergeTool::external)),
         }
     }
@@ -418,6 +426,14 @@ impl MergeEditor {
                 let tree_id = edit_merge_builtin(tree, &merge_tool_files).map_err(Box::new)?;
                 Ok((tree_id, None))
             }
+            MergeTool::Ours => {
+                let tree_id = pick_conflict_side(tree, &merge_tool_files, 0)?;
+                Ok((tree_id, None))
+            }
+            MergeTool::Theirs => {
+                let tree_id = pick_conflict_side(tree, &merge_tool_files, 1)?;
+                Ok((tree_id, None))
+            }
             MergeTool::External(editor) => external::run_mergetool_external(
                 ui,
                 &self.path_converter,
@@ -428,6 +444,21 @@ impl MergeEditor {
             ),
         }
     }
+}
+
+fn pick_conflict_side(
+    tree: &MergedTree,
+    merge_tool_files: &[MergeToolFile],
+    add_index: usize,
+) -> Result<MergedTreeId, BackendError> {
+    let mut tree_builder = MergedTreeBuilder::new(tree.id());
+    for merge_tool_file in merge_tool_files {
+        tree_builder.set_or_remove(
+            merge_tool_file.repo_path.clone(),
+            Merge::resolved(merge_tool_file.conflict.get_add(add_index).unwrap().clone()),
+        );
+    }
+    tree_builder.write_tree(tree.store())
 }
 
 #[cfg(test)]